Anxiety Data

A fictional dataset with missings to exemplify the use of HLM method for clinical significance.

Usage

anxiety

Format

A data frame with 580 rows and 4 variables:

subject

Participant

treatment

Treatment. Either Placebo or Intervention

measurement

Number of measurement

anxiety

Anxiety score (lower is better)

Details

In a fictional clinical trial, participants were split up to belong to either a medical placebo ("Placebo") or psychotherapeutic intervention ("Intervention") group.

They underwent outpatient treatment during which they were followed for 5 measurements at which a fictional anxiety score was measured. This anxiety score may range from 0 - 60.

The functional population (i.e., non-anxious individuals) can be expected to have a mean score of M = 8 points with a standard deviation of SD = 4.