Get Number Of Participants From A clinisig Object

Get Number Of Participants From A clinisig Object

Usage

get_n(x, which = "all")

Arguments

x

A clinisig object

which

Which n should be returned? Available options are

"all", n in the original and used data set (the default)

"original", n in the original dataset

"used", n in the used data set, so after conversion to wide format and omitting cases with missing values

Value

A tibble with number of participants

Examples

results <- jacobson_1989 %>%
  clinical_significance(
    id = subject,
    time = time,
    outcome = gds,
    pre = "pre",
    reliability = 0.80,
    m_functional = 30,
    sd_functional = 10,
    type = "c"
  )

get_n(results)
#> # A tibble: 1 × 3
#>   n_original n_used percent_used
#>        <int>  <int>        <dbl>
#> 1         30     26        0.867
get_n(results, which = "original")
#> # A tibble: 1 × 1
#>   n_original
#>        <int>
#> 1         30
get_n(results, which = "used")
#> # A tibble: 1 × 1
#>   n_used
#>    <int>
#> 1     26