Get A Summary Table From A cs_analysis Object

Retrieve the summary table in a tidy tibble format. This is especially useful to plot the results or conduct sensitivity analyses.

Usage

cs_get_summary(x, ...)

# Default S3 method
cs_get_summary(x, which = c("individual", "group"), ...)

# S3 method for class 'cs_anchor_group_within'
cs_get_summary(x, ...)

# S3 method for class 'cs_anchor_group_between'
cs_get_summary(x, ...)

Arguments

x

An object of class cs_analysis

...

Additional arguments passed to the respective method

which

Which level of summary table to return. This is only necessary for method "HA" since two summary tables are reported. Available are

individual, the default

group, group level results according to Hageman & Arrindell (1999)

Value

A tibble with clinical significance categories

References

Hageman, W. J., & Arrindell, W. A. (1999). Establishing clinically significant change: increment of precision and the distinction between individual and group level analysis. Behaviour Research and Therapy, 37(12), 1169–1193. https://doi.org/10.1016/S0005-7967(99)00032-7

See also

Examples

anchor_results <- claus_2020 |>
  cs_anchor(
    id,
    time,
    bdi,
    pre = 1,
    post = 4,
    mid_improvement = 8
  )

cs_get_summary(anchor_results)
#> # A tibble: 3 × 3
#>   category         n percent
#>   <fct>        <int> <chr>  
#> 1 Improved        20 50.00% 
#> 2 Unchanged       17 42.50% 
#> 3 Deteriorated     3 7.50%  

# Get summary table for a group level analysis
anchor_results_grouped <- claus_2020 |>
  cs_anchor(
    id,
    time,
    bdi,
    pre = 1,
    post = 4,
    mid_improvement = 8,
    target = "group"
  )

cs_get_summary(anchor_results_grouped)
#> # A tibble: 1 × 6
#>   difference lower upper    ci     n category                              
#>        <dbl> <dbl> <dbl> <dbl> <int> <chr>                                 
#> 1      -9.40 -12.8 -5.85  0.95    40 Probably clinically significant effect

# Get group-wise summary table for the Hageman & Arrindell method
combined_results <- claus_2020 |>
  cs_combined(
    id,
    time,
    bdi,
    pre = 1,
    post = 4,
    m_functional = 8,
    sd_functional = 8,
    reliability = 0.80,
    rci_method = "HA"
  )

cs_get_summary(combined_results)
#> # A tibble: 5 × 3
#>   category         n percent
#>   <fct>        <int> <chr>  
#> 1 Recovered        7 17.50% 
#> 2 Improved        18 45.00% 
#> 3 Unchanged       15 37.50% 
#> 4 Deteriorated     0 0.00%  
#> 5 Harmed           0 0.00%  
cs_get_summary(combined_results, which = "group")
#> # A tibble: 2 × 2
#>   category   percent
#>   <chr>        <dbl>
#> 1 Changed      0.841
#> 2 Functional   0.354